Extract Direct OTP (SMS) whitelisting to mixin
This logic is going to be shared between Direct (SMS) and Time (APP) OTP
classes.

Refactoring the current implementation on the Direct OTP class so can be
re-used in both.
